flux패턴 연습용으로 로직을 상위 컴포넌트에 뺀다.
