+++ vite-project/src/component/uploadComponent/uploadpage/UploadSubmitComponent.jsx
@@ -1,9 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import { useUploadStore } from "../../../model";
-import { postUploadApi } from "../../../api/postUploadApi";
-import { getCurrentTime } from "../../../util/uploadUtil/getCurrentTime";
-import { useNavigate } from "react-router-dom";
+
+
 
 const SubmitButton = styled.button`
   cursor: pointer;
@@ -24,32 +22,9 @@ const SubmitButton = styled.button`
   }
 `;
 
-export function UploadSubmitComponent() {
-  const navigator = useNavigate();
-  const { uploadState, file, thumbnailImage, title, userId } = useUploadStore();
-  
+export function UploadSubmitComponent({api}) {
   function handleClick() {
-    if (uploadState == true) {
-      console.log("upload");
-      postUploadApi(file, thumbnailImage, title, userId)
-      //FIXME: uplaodTime 제외하기.
-        .then((response) => {
-          if (response.status == 200) {
-            alert("Success!");
-            navigator("/");
-          } else {
-            alert(
-              "뭔가 문제가 생겼음, 에러 코드에 따른 예외처리는 하지 않음, upload api 명세를 다시 확인해보고 어떤 에러인지 확인해보시길"
-            );
-            navigator("/upload");
-          }
-        })
-        .catch((e) => {
-          console.log(e);
-        });
-    } else if (uploadState == false) {
-      alert("파일 또는 제목을 먼저 입력하세요.");
-    }
+    api();
   }
 
   return (
